backend/requirements.txt
absl-py==2.2.2
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.11.0
astunparse==1.6.3
attrs==25.3.0
certifi==2025.4.26
cffi==2.0.0
charset-normalizer==3.4.1
click==8.3.0
colorama==0.4.6
cryptography==46.0.4
dataclasses-json==0.6.7
dill==0.4.0
distlib==0.4.0
distro==1.9.0
Django==6.0.1
django-cors-headers==4.9.0
djangorestframework==3.16.1
dnspython==2.8.0
docker==7.1.0
faiss-cpu==1.13.2
fastapi==0.129.0
filelock==3.19.1
flatbuffers==25.2.10
frozenlist==1.8.0
fsspec==2026.2.0
gast==0.6.0
google-pasta==0.2.0
greenlet==3.3.1
grpcio==1.71.0
h11==0.16.0
h5py==3.13.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
huggingface_hub==1.4.1
idna==3.10
Jinja2==3.1.6
jiter==0.13.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jwt==1.4.0
keras==3.9.2
langchain==1.2.10
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.13
langchain-openai==1.1.10
langchain-text-splitters==1.1.0
langgraph==1.0.8
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.7
langgraph-sdk==0.3.6
langsmith==0.7.4
libclang==18.1.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.2
mcp==1.15.0
mdurl==0.1.2
ml_dtypes==0.5.1
mongoengine==0.29.1
mpmath==1.3.0
multidict==6.7.1
mypy_extensions==1.1.0
namex==0.0.9
networkx==3.6.1
numpy==2.1.0
openai==2.21.0
opencv-python==4.12.0.88
opt_einsum==3.4.0
optree==0.15.0
orjson==3.11.7
ormsgpack==1.12.2
packaging==25.0
pandas==2.2.2
pillow==11.3.0
platformdirs==4.4.0
propcache==0.4.1
protobuf==5.29.4
pycparser==3.0
pydantic==2.11.9
pydantic-settings==2.11.0
pydantic_core==2.33.2
Pygments==2.19.1
pymongo==4.16.0
pypdf==6.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2024.1
pywin32==311
PyYAML==6.0.3
referencing==0.36.2
regex==2026.1.15
requests==2.32.5
requests-toolbelt==1.0.0
reservoirpy==0.3.13
rich==14.0.0
rpds-py==0.27.1
safetensors==0.7.0
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==5.2.3
setuptools==80.0.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.46
sqlparse==0.5.5
sse-starlette==3.0.2
starlette==0.48.0
sympy==1.14.0
tenacity==9.1.4
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==3.0.1
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.2
torch==2.10.0
tqdm==4.67.1
transformers==5.2.0
typer==0.24.0
typer-slim==0.24.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.13.2
tzdata==2024.1
unreal==0.1.1
UnrealCV==1.1.7
urllib3==2.4.0
uuid_utils==0.14.0
uv==0.8.22
uvicorn==0.37.0
virtualenv==20.34.0
Werkzeug==0.16.1
wheel==0.45.1
wrapt==1.17.2
xgboost==3.0.0
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
